Silence compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@17947 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RevisionGraph/ExactCopyFroms.cpp
@@ -41,8 +41,8 @@ void CExactCopyFroms::Apply (CVisibleGraph* graph, CVisibleGraphNode* node)
 {
     // remove node, if it is neither "M", "A", "D" nor "R"
 
-	const DWORD forbiddenMask = CNodeClassification::IS_OPERATION_MASK
-					          | CNodeClassification::MUST_BE_PRESERVED;
+	const DWORD forbiddenMask = DWORD(CNodeClassification::IS_OPERATION_MASK
+					          | CNodeClassification::MUST_BE_PRESERVED);
 
     if (node->GetClassification().Matches (0, forbiddenMask))
     {
